requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cuid2==2.0.1
distro==1.9.0
fastapi==0.115.12
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.51
langsmith==0.3.29
openai==1.72.0
orjson==3.10.16
packaging==24.2
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.1.0
python-ulid==3.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
sentient-agent-framework==0.3.0
sniffio==1.3.1
starlette==0.46.1
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions>=4.13.2
urllib3==2.4.0
uvicorn==0.34.0
zstandard==0.23.0
beautifulsoup4==4.12.2
selenium==4.18.1
webdriver-manager==4.0.1
playwright==1.42.0
bs4==0.0.1
soupsieve==2.5
jsonschema==4.21.1 
